requirements.txt
streamlit
tensorflow-2.14.0
numpy
pillow
openai==0.28.0
python-dotenv
